now adding spaces before redir if needed and handling >< syntax error
