+++ includes/parser.h
@@ -6,7 +6,7 @@
 /*   By: mamesser <mamesser@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/07 17:20:07 by valmpani          #+#    #+#             */
-/*   Updated: 2023/08/12 16:02:19 by mamesser         ###   ########.fr       */
+/*   Updated: 2023/08/12 18:04:52 by mamesser         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
