+++ includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: mamesser <mamesser@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/06 16:13:07 by mamesser          #+#    #+#             */
-/*   Updated: 2023/08/12 14:21:12 by mamesser         ###   ########.fr       */
+/*   Updated: 2023/08/12 17:12:32 by mamesser         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,10 +23,13 @@ void	ft_cmd_lst_free(t_cmd **lst);
 void	ft_lst_free(t_lex **lst);
 
 char	**split_args(char *input);
-char	*handle_redirects(char *input, int j);
-int		calc_redir_wo_space(char *input);
-int		ft_isspace(char c);
 char	*ft_charjoin_mod(char *s1, char s2);
 char	*ft_strjoin_mod(char *s1, char *s2);
+
+char	*handle_redirects(char *input, int j);
+char	*add_space_pre_redir(char *input, int j);
+char	*add_space_post_redir(char *input, int j);
+int		calc_redir_wo_space(char *input);
+int		calc_redir_wo_space2(char *input);
 
 #endif
