+++ mandatory/utils/utils.c
@@ -6,7 +6,7 @@
 /*   By: mamesser <mamesser@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/06 15:52:47 by valmpani          #+#    #+#             */
-/*   Updated: 2023/08/12 15:26:38 by mamesser         ###   ########.fr       */
+/*   Updated: 2023/08/12 17:12:18 by mamesser         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,56 +37,6 @@ char	**split_args(char *input)
 	split = ft_split(input, 26);
 	free(input); // add new variable in the main for copy
 	return (split);
-}
-
-char	*handle_redirects(char *input, int j)
-{
-	int		i;
-	char	*new_input;
-	int		red_no_sp;
-
-	i = -1;
-	red_no_sp = calc_redir_wo_space(input);
-	new_input = ft_calloc((ft_strlen(input) + red_no_sp + 1), sizeof(*new_input));
-	if (!new_input)
-		return (NULL);
-	while (input[++i])
-	{
-		new_input[j++] = input[i];
-		if ((input[i] == '<' || input[i] == '>')
-			&& (input[i + 1] != '<' && input[i + 1] != '>'
-				&& !(input[i + 1] == 26) && input[i + 1]))
-		{
-			new_input[j++] = 26;
-			if (i > 0 && (input[i] == '<' && input[i - 1] == '>')
-				|| (input[i] == '>' && input[i - 1] == '<'))
-				j--; //will be done in lexer
-		}
-	}
-	// free(input);
-	return (new_input);
-}
-
-int	calc_redir_wo_space(char *input)
-{
-	int	redir_no_sp;
-	int	i;
-
-	redir_no_sp = 0;
-	i = -1;
-	while (input[++i + 1])
-	{
-		if ((input[i] == '<' || input[i] == '>')
-			&& (input[i + 1] != '<' && input[i + 1] != '>'
-				&& !(input[i + 1] == 26) && input[i + 1]))
-		{
-			redir_no_sp++;
-			if (i > 0 && (input[i] == '<' && input[i - 1] == '>')
-				|| (input[i] == '>' && input[i - 1] == '<'))
-				redir_no_sp--;
-		}
-	}
-	return (redir_no_sp);
 }
 
 char	*ft_charjoin_mod(char *s1, char s2)
